chore: refactor (#48)
* chore: refactor

* Update e-token/src/processor/shuttle_close_schedule.rs


* chore: refactor

---------
