+++ e-token/src/processor/mod.rs
@@ -21,10 +21,10 @@ pub mod merge_shuttle_into_ephemeral_ata;
 pub mod process_transfer_queue_tick;
 pub mod rent_pda;
 pub mod reset_ephemeral_ata_permission;
-pub mod undelegate_and_close_shuttle_ephemeral_ata;
+pub(crate) mod shuttle_close_schedule;
+pub mod undelegate_and_close_shuttle_to_owner;
 pub mod undelegate_ephemeral_ata;
 pub mod undelegate_ephemeral_ata_permission;
-pub mod undelegate_withdraw_and_close_shuttle_ephemeral_ata;
 pub mod undelegation_callback;
 pub(crate) mod utils;
 pub mod withdraw_spl_tokens;
@@ -56,10 +56,9 @@ pub use merge_shuttle_into_ephemeral_ata::process_merge_shuttle_into_ephemeral_a
 pub use process_transfer_queue_tick::process_transfer_queue_tick;
 pub use rent_pda::process_initialize_rent_pda;
 pub use reset_ephemeral_ata_permission::process_reset_ephemeral_ata_permission;
-pub use undelegate_and_close_shuttle_ephemeral_ata::process_undelegate_and_close_shuttle_ephemeral_ata;
+pub use undelegate_and_close_shuttle_to_owner::process_undelegate_and_close_shuttle_to_owner;
 pub use undelegate_ephemeral_ata::process_undelegate_ephemeral_ata;
 pub use undelegate_ephemeral_ata_permission::process_undelegate_ephemeral_ata_permission;
-pub use undelegate_withdraw_and_close_shuttle_ephemeral_ata::process_undelegate_withdraw_and_close_shuttle_ephemeral_ata;
 pub use undelegation_callback::process_undelegation_callback;
 pub use withdraw_spl_tokens::process_withdraw_spl_tokens;
 pub use withdraw_through_delegated_shuttle_with_merge::process_withdraw_through_delegated_shuttle_with_merge;
